backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
SQLAlchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
aiofiles==23.2.1
openpyxl==3.1.5
pytesseract==0.3.13
Pillow==10.4.0
reportlab==4.2.2
